pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-harness"
version = "0.1.0"
description = "Multi-agent orchestration harness with true isolation for the Agent Protocol"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = [
    "ai",
    "agents",
    "claude",
    "orchestration",
    "multi-agent",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "anthropic>=0.18.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "watchdog>=3.0.0",
    "mcp>=0.1.0",
]

[project.optional-dependencies]
redis = [
    "redis>=5.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
all = [
    "agent-harness[redis,dev]",
]

[project.scripts]
agent-harness = "agent_harness.cli:main"
agent-harness-mcp = "agent_harness.mcp_server:main"

[project.urls]
Homepage = "https://github.com/yourname/agent-harness"
Documentation = "https://github.com/yourname/agent-harness#readme"
Repository = "https://github.com/yourname/agent-harness"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_harness"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
